Added notifications for summaries

+++ app/controllers/notifications_controller.rb
@@ -1,5 +1,6 @@
 class NotificationsController < ApplicationController
-  before_action :logged_in_user, only: [:index, :important, :delete, :timeline]
+  before_action :logged_in_user, only: [:index, :important, :delete, :delete_all,
+                       :summary_selection, :selection, :timeline, :reference, :comment]
 
   def index
     if params[:filter]
@@ -26,27 +27,35 @@ class NotificationsController < ApplicationController
     @notification = Notification.new
     case @filter
       when :timeline
-        timeline_ids = NotificationTimeline.where( user_id: current_user.id, read: false ).pluck( :timeline_id )
+        timeline_ids = NotificationTimeline.where( user_id: current_user.id).pluck( :timeline_id )
         @timelines = Timeline.select(:id, :name).where( id: timeline_ids ).page(params[:page]).per(20)
+      when :summary
+        summary_ids = NotificationSummary.where( user_id: current_user.id ).pluck( :summary_id )
+        @summaries = Summary.select(:id, :timeline_id,
+                                    :content).where( id: summary_ids ).page(params[:page]).per(20)
       when :reference
-        reference_ids = NotificationReference.where( user_id: current_user.id, read: false ).pluck( :reference_id )
+        reference_ids = NotificationReference.where( user_id: current_user.id).pluck( :reference_id )
         @references = Reference.select(:id, :timeline_id, :title_fr).where( id: reference_ids ).page(params[:page]).per(20)
       when :comment
-        comment_ids = NotificationComment.where( user_id: current_user.id, read: false ).pluck( :comment_id )
+        comment_ids = NotificationComment.where( user_id: current_user.id ).pluck( :comment_id )
         @comments = Comment.select(:id, :timeline_id, :reference_id,
                                :f_1_content).where( id: comment_ids ).page(params[:page]).per(20)
       when :selection
-        comment_sel_ids = NotificationSelection.where( user_id: current_user.id, read: false ).pluck( :new_comment_id )
+        comment_sel_ids = NotificationSelection.where( user_id: current_user.id).pluck( :new_comment_id )
         @selections = Comment.select(:id, :timeline_id, :reference_id,
                                      :f_1_content).where( id: comment_sel_ids ).page(params[:page]).per(20)
+      when :summary_selection
+        summary_sel_ids = NotificationSummarySelection.where( user_id: current_user.id).pluck( :new_summary_id )
+        @summary_selections = Summary.select(:id, :timeline_id,
+                                     :content).where( id: summary_sel_ids ).page(params[:page]).per(20)
     end
   end
 
   def important
-    win_ids = NotificationSelectionWin.where( user_id: current_user.id, read: false ).pluck( :comment_id )
+    win_ids = NotificationSelectionWin.where( user_id: current_user.id ).pluck( :comment_id )
     @wins = Comment.select(:id, :timeline_id, :reference_id,
                            :f_1_content).where( id: win_ids ).page(params[:page]).per(20)
-    loss_ids = NotificationSelectionLoss.where( user_id: current_user.id, read: false ).pluck( :comment_id )
+    loss_ids = NotificationSelectionLoss.where( user_id: current_user.id ).pluck( :comment_id )
     @losses = Comment.select(:id, :timeline_id, :reference_id,
                              :f_1_content).where( id: loss_ids ).page(params[:page]).per(20)
   end
@@ -58,6 +67,13 @@ class NotificationsController < ApplicationController
                               timeline_id: params[:notification][:timeline_ids] )
         notifs.destroy_all
         redirect_to notifications_index_path( filter: :timeline)
+        return
+      end
+      if params[:notification][:summary_ids]
+        notifs = NotificationSummary.where( user_id: current_user.id,
+                                            summary_id: params[:notification][:summary_ids] )
+        notifs.destroy_all
+        redirect_to notifications_index_path( filter: :summary)
         return
       end
       if params[:notification][:reference_ids]
@@ -81,6 +97,13 @@ class NotificationsController < ApplicationController
         redirect_to notifications_index_path( filter: :selection)
         return
       end
+      if params[:notification][:sel_summary_ids]
+        notifs = NotificationSummarySelection.where( user_id: current_user.id,
+                                              new_summary_id: params[:notification][:sel_summary_ids] )
+        notifs.destroy_all
+        redirect_to notifications_index_path( filter: :summary_selection)
+        return
+      end
     end
     redirect_to notifications_index_path
   end
@@ -90,20 +113,23 @@ class NotificationsController < ApplicationController
       when :timeline.to_s
         notifs = NotificationTimeline.where( user_id: current_user.id )
         notifs.destroy_all
-        redirect_to notifications_index_path
+      when :summary.to_s
+        notifs = NotificationSummary.where( user_id: current_user.id )
+        notifs.destroy_all
       when :reference.to_s
         notifs = NotificationReference.where( user_id: current_user.id )
         notifs.destroy_all
-        redirect_to notifications_index_path
       when :comment.to_s
         notifs = NotificationComment.where( user_id: current_user.id )
         notifs.destroy_all
-        redirect_to notifications_index_path
       when :selection.to_s
         notifs = NotificationSelection.where( user_id: current_user.id )
         notifs.destroy_all
-        redirect_to notifications_index_path
+      when :summary_selection.to_s
+        notifs = NotificationSummarySelection.where( user_id: current_user.id )
+        notifs.destroy_all
     end
+    redirect_to notifications_index_path
   end
 
   def timeline
@@ -111,6 +137,13 @@ class NotificationsController < ApplicationController
                         timeline_id: notification_params )
     notif.destroy
     redirect_to timeline_path( notification_params )
+  end
+
+  def summary
+    notif = NotificationSummary.find_by( user_id: current_user.id,
+                                           summary_id: notification_params )
+    notif.destroy
+    redirect_to summary_path( notification_params )
   end
 
   def reference
@@ -132,6 +165,13 @@ class NotificationsController < ApplicationController
                                           new_comment_id: notification_params )
     notif.destroy
     redirect_to comment_path( notification_params )
+  end
+
+  def summary_selection
+    notif = NotificationSummarySelection.find_by( user_id: current_user.id,
+                                           new_summary_id: notification_params )
+    notif.destroy
+    redirect_to summary_path( notification_params )
   end
 
   private
